OT249-25-Migrate to UserDto.
