+++ src/main/java/com/alkemy/ong/dto/UserDto.java
@@ -4,8 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
 import java.io.Serializable;
+import java.util.Collection;
+import java.util.List;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
@@ -14,7 +18,7 @@ import javax.validation.constraints.NotBlank;
 @NoArgsConstructor
 @Setter
 @Getter
-public class UserDto implements Serializable {
+public class UserDto implements Serializable, UserDetails {
 
 	private static final long serialVersionUID = 1L;
 
@@ -35,7 +39,44 @@ public class UserDto implements Serializable {
 	private String photo;
 	
 	@NotBlank
-	private String rol;
+	private RoleDto role;
 
-	
+	private boolean deleted;
+
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+
+		return List.of( this.getRole() );
+	}
+
+
+	@Override
+	public String getUsername() {
+
+		return this.getEmail();
+	}
+
+	@Override
+	public boolean isAccountNonExpired() {
+
+		return !this.deleted;
+	}
+
+	@Override
+	public boolean isAccountNonLocked() {
+
+		return !this.deleted;
+	}
+
+	@Override
+	public boolean isCredentialsNonExpired() {
+
+		return !this.deleted;
+	}
+
+	@Override
+	public boolean isEnabled() {
+
+		return !this.deleted;
+	}
 }
