+++ src/main/java/com/alkemy/ong/security/CustomExceptionHandler.java
@@ -2,8 +2,11 @@ package com.alkemy.ong.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
+import org.springframework.web.servlet.HandlerExceptionResolver;
 
 import javax.servlet.FilterChain;
 import javax.servlet.http.HttpServletRequest;
@@ -14,12 +17,9 @@ import java.io.IOException;
 @Component
 public class CustomExceptionHandler extends OncePerRequestFilter {
 
-    //  @Autowired
-    //  @Qualifier("handlerExceptionResolver")
-    // private HandlerExceptionResolver resolver;
-
-//    @Override
-//    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) {
+    @Autowired
+    @Qualifier("handlerExceptionResolver")
+    private HandlerExceptionResolver resolver;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) {
@@ -28,7 +28,7 @@ public class CustomExceptionHandler extends OncePerRequestFilter {
             filterChain.doFilter(request, response);
         } catch (Exception e) {
             log.error("Spring Security Filter Chain Exception:", e);
-            ///  resolver.resolveException(request, response, null, e);
+            resolver.resolveException(request, response, null, e);
             try {
                 new ObjectMapper().writeValue(response.getOutputStream(), e);
             } catch (IOException ex) {
